# server/design/openapi.yaml
openapi: 3.0.3

info:
  title: Astitva API
  version: 0.9.0
  description: REST API for local Dev and Stage profiles or the Render Production profile.

servers:
  - url: http://localhost:3001
    description: Local Express server
  - url: https://{service}.onrender.com
    description: Render Production server
    variables:
      service:
        default: SERVICE_NAME

tags:
  - name: System
  - name: Authentication
  - name: Finance
  - name: Daily Priorities

paths:
  /api/priorities/days/{date}:
    $ref: './priorities.openapi.json#/paths/~1api~1priorities~1days~1{date}'
  /api/priorities/days/{date}/priorities:
    $ref: './priorities.openapi.json#/paths/~1api~1priorities~1days~1{date}~1priorities'
  /api/priorities/days/{date}/priorities/{priorityId}:
    $ref: './priorities.openapi.json#/paths/~1api~1priorities~1days~1{date}~1priorities~1{priorityId}'
  /api/finance/summary:
    $ref: './finance.openapi.json#/paths/~1api~1finance~1summary'
  /api/finance/connections:
    $ref: './finance.openapi.json#/paths/~1api~1finance~1connections'
  /api/finance/connection-sessions:
    $ref: './finance.openapi.json#/paths/~1api~1finance~1connection-sessions'
  /api/finance/sync:
    $ref: './finance.openapi.json#/paths/~1api~1finance~1sync'
  /api/finance/accounts/{accountId}:
    $ref: './finance.openapi.json#/paths/~1api~1finance~1accounts~1{accountId}'
  /api/finance/connections/{connectionId}:
    $ref: './finance.openapi.json#/paths/~1api~1finance~1connections~1{connectionId}'
  /api/diet/days/{date}:
    $ref: './diet.openapi.json#/paths/~1api~1diet~1days~1{date}'
  /api/diet/meals:
    $ref: './diet.openapi.json#/paths/~1api~1diet~1meals'
  /api/diet/meals/{mealId}:
    $ref: './diet.openapi.json#/paths/~1api~1diet~1meals~1{mealId}'
  /api/diet/targets:
    $ref: './diet.openapi.json#/paths/~1api~1diet~1targets'
  /api/health:
    get:
      tags: [System]
      summary: Check server health
      operationId: getHealth
      responses:
        "200":
          description: The server is running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/auth/signup:
    post:
      tags: [Authentication]
      summary: Create an unverified account and send a verification email
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created. Email delivery success is reported separately.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Production signup requires an invited email address.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: An account with the email already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/auth/verify-email:
    post:
      tags: [Authentication]
      summary: Verify an email address with a single-use token
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Email verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: The email is already verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: The token is invalid, expired, or already used.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/resend-verification:
    post:
      tags: [Authentication]
      summary: Replace and resend the logged-in user's verification token
      operationId: resendVerification
      security:
        - sessionCookie: []
      responses:
        "202":
          description: A replacement verification email was accepted for delivery.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The email is already verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Local email delivery failed and the account remains unverified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Log in and create a session
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded. The response sets an HTTP-only session cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get the logged-in user
      operationId: getCurrentUser
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current user details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password-reset email without revealing account eligibility
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "202":
          description: Generic response returned for eligible, unknown, unverified, and rate-limited accounts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Replace a password using a single-use reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password changed and all active sessions invalidated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "410":
          description: The token is invalid, expired, already used, or belongs to an unverified account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Log out and delete the current session
      operationId: logout
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Logout completed.

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: astitva_dev_session
      description: Dev, Stage, and Production use isolated cookie names. Production uses Secure and SameSite=None.

  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: The Production request rate limit was exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    HealthResponse:
      type: object
      required: [status, message, environment, dataLocation, financeProvider]
      properties:
        status:
          type: string
          example: ok
        message:
          type: string
          example: Astitva server is running
        environment:
          type: string
          enum: [dev, stage, production]
        dataLocation:
          type: string
          enum: [local, cloud]
        financeProvider:
          type: object
          required: [name, environment, enabled, configured]
          properties:
            name:
              type: string
              enum: [plaid]
            environment:
              type: string
              enum: [sandbox, production]
            enabled:
              type: boolean
            configured:
              type: boolean

    SignupRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          maxLength: 80
          example: Astitva User
        email:
          type: string
          format: email
          maxLength: 254
          example: developer@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: local-demo-password

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: developer@example.com
        password:
          type: string
          format: password
          example: local-demo-password

    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: developer@example.com

    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          pattern: "^[a-fA-F0-9]{64}$"
        password:
          type: string
          format: password
          minLength: 8

    User:
      type: object
      required: [id, name, email, emailVerified]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean

    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          pattern: "^[a-fA-F0-9]{64}$"

    SignupResponse:
      type: object
      required: [user, message, emailDelivered]
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
        emailDelivered:
          type: boolean

    VerificationResponse:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string